fix: content type in context.call
in context.call, if the user provided content-type, we were sending it like , but QStash ignores this if there is a  header.

since batchJSON added the content-type: application/json, we were unable to change the content-type. In the new version, we use batch instead of batchJSON and use the headers provided in the step as content-type if they are given.