Refactor internal steps, headers and body creation (#87)
* fix: refactor step submissions

* fix: bump ai dependencies and fix tests

* fix: refactor getHeaders method

* fix: tests

* fix: remove comment

* fix: fmt

* ci: fix wrangler publish

* feat: fix callback-failurecallback headers

handles https://linear.app/upstash/issue/DX-1718/workflow-js-fix-fromcallback-headers

* fix: tests

* fix: review

* fix: rm unused invokeWorkflow

* fix: test

* fix: express dependency

we were getting type errors after @types/express v5.0.1 was released a week ago. Pinning the version of this package in the workflow-js and examples/express to 5.0.0 fixed the issue. But this wasn't a good solution in the long term.

another solution was removing the dependency from examples/express and keeping it in workflow-js. We don't import anything from @types/express but we still need it in our dependencies.

* fix: add debug logs