fix: use process.env if context.env isn't defined in hono
It looks like context.env is only defined in Cloudflare Workers. https://hono.dev/docs/api/context#env

When the serve method from hono is used in vercel, context.env is undefined which causes an error.

Fixed the issue by checking if context.env is defined and falling back to process.env if it isn't.

Fixes: https://github.com/upstash/workflow-js/issues/91